Fixed some bugs, partially by Oscar Lesta.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@13767 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/drivers/poke.h
@@ -41,7 +41,7 @@ typedef struct {
 	uint32	signature;
 	uint16	port;
 	uint8	size;		// == index for POKE_PORT_INDEXED_*
-	uint16	value;
+	uint32	value;
 } port_io_args;
